[SandboxIR][NFC] Move Region from SandboxVectorizer to SandboxIR.
I'm planning to add RegionPass and RegionPassManager next to the
equivalent FunctionPass and FunctionPassManager in SandboxIR, which
means that SandboxIR has to kwow about Regions.

There's nothing vectorizer-specific about the Region class, so this is
a straightforward file move.
